Add persistence for user engagement daily

+++ schema/user_engagement_daily.sql
@@ -2,8 +2,8 @@ CREATE TABLE user_engagement_daily (
   user_id       VARCHAR(64),
   date          DATE,
 
-  messages_sent INT,
-  active_channels INT,
+  messages_sent LONG,
+  active_channels LONG,
 
   PRIMARY KEY (user_id, date)
 );
